fix(fish): Improve fix_broken_services_by_nixos

+++ fish/functions/fix_broken_services_by_nixos.fish
@@ -1,27 +1,30 @@
 # Reinstall systemd user services broken by NixOS updates
 function fix_broken_services_by_nixos
-    set services $(fd service ~/.config/systemd/user/)
-    if test $(count $services) -eq 0
-        # Hardcoded services in case they were deleted
-        set names gammastep swayidle xdg-desktop-portal-rewrite-launchers
-        set services
+    # Retrieve all systemd user services in ~/.config/systemd/user/
+    set services (fd '.service' ~/.config/systemd/user/)
+
+    # If no services are found, use a predefined list of service names
+    if test (count $services) -eq 0
+        set -l names gammastep swayidle xdg-desktop-portal-rewrite-launchers
         for name in $names
             set services $services ~/.config/systemd/user/$name.service
         end
     end
+
+    # Repair each broken service
     for service_path in $services
-        set -l nix_path $(realpath --canonicalize-missing $service_path)
-        if not test -e $nix_path
+        set -l nix_path (realpath --canonicalize-missing $service_path)
+
+        # Check if service path is broken (missing target)
+        if test -L $service_path -a ! -e $nix_path
             set_color --bold red
-            echo "Fixing $(basename $service_path) broken by a NixOS update..."
+            echo "Fixing (basename $service_path) broken by a NixOS update..."
             set_color normal
-            if test -L $service_path
-                rm $service_path 2>/dev/null
-            end
-            systemctl enable --user $(basename $service_path) 2>/dev/null
-            systemctl restart --user $(basename $service_path) 2>/dev/null
-        else
-            break # All broken at the same time
+
+            # Remove broken symlink and re-enable the service
+            rm $service_path 2>/dev/null
+            systemctl enable --user (basename $service_path) 2>/dev/null
+            systemctl restart --user (basename $service_path) 2>/dev/null
         end
     end
 end
